Beta xml now emitting correct edn and processing in batches
The previous iteration was not correctly collecting the beta XML; only
the first file was being read on any given operation. This problem has
been addressed. Additionally the file write operation now occurs on
the main thread, which should eliminate the race condition that was
previously a problem when the main XML file closed before writes were
complete. The output is now a set of batched files, which makes it
easier to emit correct EDN.